Make client interface 64 bits wide on Sparc64
git-svn-id: svn://coreboot.org/openbios/openbios-devel@164 f158a5a8-5612-0410-a976-696ce0be7e32

+++ modules/client.c
@@ -28,8 +28,8 @@
 
 typedef struct prom_args {
         const char 	*service;
-        int 		nargs;
-        int 		nret;
+        long 		nargs;
+        long 		nret;
         ulong 		args[10];		/* MAX NUM ARGS! */
 } prom_args_t;
